+++ sandbox/libs/authn/src/main/java/org/opensearch/authn/jwt/package-info.java
@@ -1,7 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-/** Classes for the JWT related functionality */
-package org.opensearch.authn.jwt;