+++ sandbox/libs/authn/src/main/java/org/opensearch/authn/package-info.java
@@ -3,5 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/** Classes for the internal OpenSearch authentication */
+/** A suite of convenient APIs for interacting with security.
+ *  Contains {@link org.opensearch.authn.AuthenticationManager} which provides methods for interacting with a subject
+ *  {@link org.opensearch.authn.Subject} provides methods for interrogating a subject like `isAuthenticated()` and
+ *  `checkPermission` to check if a currently authenticated subject has permissions to perform an action
+ */
 package org.opensearch.authn;